Move inverse functions to inv.rs
