+++ eyelid-match-ops/src/primitives/poly/modular_poly.rs
@@ -23,6 +23,7 @@ use crate::primitives::poly::{mod_poly, mul_poly, new_unreduced_poly_modulus_slo
 
 pub(super) mod modulus;
 pub(super) mod mul;
+pub(crate) mod inv;
 
 mod trivial;
