+++ eyelid-match-ops/src/primitives/yashe.rs
@@ -1,7 +1,13 @@
 //! Implementation of YASHE cryptosystem
 //! `<https://eprint.iacr.org/2013/075.pdf>`
 
-use crate::primitives::poly::{inverse, modular_poly::Poly, Coeff};
+use crate::primitives::poly::{
+    modular_poly::{
+        inv::inverse,
+        Poly,
+    },
+    Coeff,
+};
 use ark_ff::{One, UniformRand};
 use rand::rngs::ThreadRng;
 use rand_distr::{Distribution, Normal};
