+++ eyelid-match-ops/src/primitives/yashe/test/keygen.rs
@@ -1,8 +1,11 @@
 //! Unit tests for Key Generation
 
 use crate::primitives::{
-    poly::FULL_RES_POLY_DEGREE,
-    yashe::{inverse, Coeff, Poly, Yashe, YasheParams},
+    poly::{
+        FULL_RES_POLY_DEGREE,
+        modular_poly::inv::inverse,
+    },
+    yashe::{Coeff, Poly, Yashe, YasheParams},
 };
 use ark_ff::One;
 use ark_poly::Polynomial;
